修复兼容问题

+++ backend/run_dpsk_ocr_pdf.py
@@ -52,6 +52,8 @@ else:
 print(f"=" * 60)
 
 print("🚀 正在初始化 LLM 模型（这可能需要一些时间）...")
+# Tesla T4 (compute capability 7.5) 不支持 bfloat16，需要使用 float16
+import torch
 llm = LLM(
     model=MODEL_PATH,
     hf_overrides={"architectures": ["DeepseekOCRForCausalLM"]},
@@ -63,7 +65,8 @@ llm = LLM(
     max_num_seqs=MAX_CONCURRENCY,
     tensor_parallel_size=1,
     gpu_memory_utilization=0.9,
-    disable_mm_preprocessor_cache=True
+    disable_mm_preprocessor_cache=True,
+    dtype=torch.float16,  # 使用 float16 而不是 bfloat16，以支持 Tesla T4
 )
 print("✅ LLM 模型初始化完成")
